Move clients info into a separate module
Logically better and it contains `next_update_index` that shouldn't be
used or changed outside.
